Publish SNS events when passing search results to websocket server
It is intended to refactor the flow of search results back to the client by passing through redis pubsub to ensure that events are received by the socket server instance that the browser client is connected to. In order to commence this search results will be passed via SNS to a lambda pushing them to redis pubsub.

Once the websocket server is correctly reading search results from redis the direct http POST will be removed.