Remove items with n url prop when saving to S3 on publish
We want to be able to send items back into the results pipe to the client without also necessarily saving them to s3 as well. In particular, it does not make sense to save items to s3 which do not url.

As such, we remove items from the items array which have no url property before saving to s3.

Also slightly refactor the tests for this method to more explicitly define the calls to downstream AWS services by stubbing the AWS SDK.