Ensure that only the results that are in S3 are pushed to client
This fixes the issue where some items could fail to be uploaded
to S3 but the errors would be ignored (and an error would be returned
only if the last upload operation failed). This could lead to the
client receiving search results that didn't have a corresponding
object in S3.