Allow `pushResultToClient` to send events with no items
There are occasions when we need to be able to send messages from providers back to the UI that do not have search results attached. For example "search complete" events. As such we need to stop the function from throwing an error if no `items` are sent.

Instead allow the SNS event to continue to be emitted, even if there are no `items` provided.